first upload to git of mini project
Mini project -
Create main menu screen in VR.
Scene should have the look and feel (UX) of the final escape room,
and the user should be introduced to the escape room through the
main menu screen.
The player should be able to start the escape room via physical or
remote UI.
Scene should have a leaderboard which will hold three top users
and show the time it took them to finish the escape room. (can be
mock up data for mini project).